requirements.txt
anyio==3.5.0
asgiref==3.5.0
cffi==1.15.0
click==8.1.2
cryptography==36.0.2
fastapi==0.75.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
pycparser==2.21
pydantic==1.9.0
PyMySQL==1.0.2
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
loginradius-v2==11.3.0